Implemented state - Added use state - migrated mockData to top level
