+++ src/app/components/eventsList/eventsList.tsx
@@ -1,24 +1,12 @@
 import styles from "./eventsList.module.scss";
 import Event from '../event/event';
+import EventInfo from "../../models/event";
 
-export default function EventsList(){
-    const mockData = [{
-        startTime: new Date("December 25, 2024 8:30:00"),
-        endTime: new Date("December 25, 2024 10:30:00"),
-        title: "Event 1"
-    }, {
-        startTime: new Date("December 25, 2024 11:30:00"),
-        endTime: new Date("December 25, 2024 12:30:00"),
-        title: "Event 2"
-    }, {
-        startTime: new Date("December 25, 2024 13:30:00"),
-        endTime: new Date("December 25, 2024 15:30:00"),
-        title: "Event 3"
-    }]
-    const totalEventsTime = mockData.reduce((accumulator, currentValue) => {
+export default function EventsList({ events }:{events: EventInfo[]}){
+    const totalEventsTime = events.reduce((accumulator, currentValue) => {
         return accumulator + ((currentValue.endTime.getTime()/1000) - (currentValue.startTime.getTime()/1000))
     }, 0);
     return(<ul id={styles.events}>{
-        mockData.map((event, index) => <Event eventInfo={event} isLastElement={index === mockData.length - 1} totalEventsSeconds={totalEventsTime}/>)
+        events.map((event, index) => <Event eventInfo={event} isLastElement={index === events.length - 1} totalEventsSeconds={totalEventsTime}/>)
     }</ul>)
 }
