+++ src/app/models/event.tsx
@@ -1,0 +1,5 @@
+export default interface EventInfo {
+    startTime: Date,
+    endTime: Date,
+    title: string
+}
